FFI-101: Define Shared Memory Ring Buffer ABI (#6)
* FFI-101: Define Shared Memory Ring Buffer ABI and Fix Build

- Established `struct RingBufferControl` in C++ and `FFI.RingBuffer.Types` in Haskell.
- Enforced 64-byte alignment and atomic offsets via `static_assert` and property tests.
- Fixed pre-existing build errors in `Control.UI` (GLUT callbacks), `SignalProcessing` (ambiguous types), and `Hardware.Ingestion` (baud rate).
- Updated Cabal configuration to support C++11 and include new FFI modules.
- Verified with `cabal test` and manual C++ compilation check.

* FFI-101: Define Shared Memory Ring Buffer ABI and Fix Build

- Established `struct RingBufferControl` in C++ and `FFI.RingBuffer.Types` in Haskell.
- Enforced 64-byte alignment and atomic offsets via `static_assert` and property tests.
- Fixed pre-existing build errors in `Control.UI` (GLUT callbacks), `SignalProcessing` (ambiguous types), and `Hardware.Ingestion` (baud rate).
- Updated Cabal configuration to support C++11 and include new FFI modules.
- Verified with `cabal test` and manual C++ compilation check.

* Update CI to use haskell:9.8

The `haskell:9.4` image relies on Debian Buster repositories which are no longer available, causing `apt-get update` to fail.
This commit updates the CI workflow to use `haskell:9.8`, which uses a newer, supported Debian base image.

* CI: Add OpenGL and GLUT dependencies to build workflow

Install `libgl1-mesa-dev`, `libglu1-mesa-dev`, and `freeglut3-dev` in the
Github Actions workflow. These C libraries are required for building
Haskell OpenGL bindings (OpenGLRaw, GLUT) used by the project.

This resolves the build failure: "Missing dependency on a foreign library: Missing (or bad) C library: GL".

* 🛠️ Forge: Centralize and Robustify Environment Setup (#9)

* feat: centralize dependency management in setup_env.sh

- Create scripts/setup_env.sh to manage system dependencies robustly.
- Detect OS (Linux/macOS) and install appropriate packages.
- Add missing dependencies (freeglut3-dev, etc.) required by Cabal.
- Update Dockerfile and CI workflow to use this script.
- Ensure set -euo pipefail and logging for better debugging.

* feat: centralize dependency management in setup_env.sh

- Create scripts/setup_env.sh to manage system dependencies robustly.
- Detect OS (Linux/macOS) and install appropriate packages.
- Add missing dependencies (freeglut3-dev, libx11-dev, etc.) required by Cabal.
- Update Dockerfile and CI workflow to use this script.
- Ensure set -euo pipefail and logging for better debugging.

---------


* Update src/FFI/RingBuffer/Types.hs

adding documentation


* Update src/FFI/RingBuffer/Types.hs

adding comments


---------

+++ src/SignalProcessing/Regression.hs
@@ -13,9 +13,9 @@ import Numeric.LinearAlgebra
 createDesignMatrix :: Vector R -> Matrix R
 createDesignMatrix xVec = fromColumns [ ones
                                       , xVec
-                                      , xVec ^ 2
-                                      , xVec ^ 3
-                                      , xVec ^ 4
+                                      , xVec ^ (2::Int)
+                                      , xVec ^ (3::Int)
+                                      , xVec ^ (4::Int)
                                       ]
   where
     n    = size xVec
@@ -24,7 +24,7 @@ createDesignMatrix xVec = fromColumns [ ones
 -- | Construct the Design Matrix for "Strict" Bi-Quadratic
 -- This forces odd coefficients to be 0 (y = ax^4 + bx^2 + c)
 createStrictBiQuadraticMatrix :: Vector R -> Matrix R
-createStrictBiQuadraticMatrix xVec = fromColumns [ ones, xVec ^ 2, xVec ^ 4 ]
+createStrictBiQuadraticMatrix xVec = fromColumns [ ones, xVec ^ (2::Int), xVec ^ (4::Int) ]
   where
     n    = size xVec
     ones = n |> repeat 1.0
